Sort out constrained primitives
This change removes double patterns and double constrained primitives.
Notably, `MIME_typed` goes away and we introduce `Content_type`
according to the book.

Furthermore, we set the correct references in the book (according to the
latest version -- though these are always in flux).

Finally, we make `Value_data_type` a string to reflect that it can
contain *any* value of the XML basic types.