Re-order V3RC02 classes to match V3RC01
We re-order the classes in V3RC02 to make the diff from the RC01 as
minimal as possible.

Additionally, we make all the minor changes to sync up with the current
working version of the meta-model. Note, however, that the working
version is still unstable, so many more changes are expected in the
future.