Introduce `get_*_or_default` as methods to V3rc2
In #70, we removed the assignments to default values in the constructors
in order to be able to test for round-trip conversions free of side
effects. This makes the life of the clients a bit more cumbersome as it
forces them to litter their code with countless ``if X is null then
default value`` checks.

In this patch, we introduce for each property with a default value a
getter method which returns the default value if the property is null.

We currently implemented this with implementation-specific methods. This
is admittedly error-prone and does not conform with the idea that most
code should be automatically generated. Unfortunately, we lack time at
the moment as we focus on the release of the schemas and C# SDK which,
on their part, need automatically generated test data.

When the time permits, we should come back and make these methods
transpilable at least.