Introduce constants
This is a major change. We introduce constants in the meta-model
including constant primitive values, constant sets of primitive values
as well as constant sets of enumeration literals.

The continuous integration is expected to fail as we change
aas-core-codegen in parallel, and this change is incompatible with the
current main branch of aas-core-codegen.