Embed data specifications in V3RC02
Concerning the data specifications, we embed them within
`Has_data_specification` instead of referencing them *via* a global
reference. The working group decided to change the rules for
serialization *after* the book was published. The data specifications
are critical in applications, but there is no possibility to access
them through a data channel as they are not part of an environment.