Enforce that lists have min. length 1 in V3RC02
We add an assertion in the tests to check that we infer the minimum
length for all the lists to be 1 except for lists where the end user
defines the semantic (*i.e.*, `Submodel_element_list` and
`Submodel_element_collection`).

This change is necessary so that we avoid ambiguity for the end users
since the distinction in semantics of a null list and an empty list is
not well-defined.