Describe all invariants in V3RC02
We describe all the invariants in V3RC02 with a human-readable text. In
consequence, this gives us human-readable messages in the SDKs. Such
messages are important when the programs are used by non-programmers,
such as GUI editors.

At this point, we also change all ``> 0`` to ``>= 1`` in invariants, so
that the formalized condition matches better the invariant description.