Re-introduce the pattern for date/time stamps UTC
We removed the invariant on the pattern since it is doubly checked in
the ``matches_xs_date_time_stamp_utc`` and
``is_xs_date_time_stamp_utc``. However, this was a mistake as we now can
not infer the pattern for the schemas and test generation.

Therefore, we re-introduce the invariant.