Enforce min. length on submodel element value (#166)
Originally, we left out the constraint on
``Submodel_element_list.value`` and
``Submodel_element_collection.value`` related to the length as we wanted
to allow the distinction between null lists and empty lists, which the
user could define per semantic ID.

However, this is problematic as many users are not aware of the
distinction. Therefore, we keep consistency with all the other lists in
the meta-model.