Sort out references
Previously, due to ignorance, we annotated references with different
prefixes (attributes with ``~`` and classes and constants with ``.``).
Eventually, we learned more about Sphinx and how it deals with
references (see [1]).

In this change, we sort out the referencing and make it a bit clearer:

* attributes, classes and constants must be references as-are,
* appearance prefixes ``!`` and ``~`` are ignored (as target languages
  usually do not work like Sphinx), and
* relaxed qualified names (references prefixed with a ``.``) are
  disallowed as we can only resolve references *within* the meta-model
  and not outside it.

This change is harmonized with the respective change in
[aas-core-codegen 46ddb38f].

[1]: https://www.sphinx-doc.org/en/master/usage/restructuredtext/roles.html#xref-syntax
[aas-core-codegen 46ddb38f]: https://github.com/aas-core-works/aas-core-codegen/commit/46ddb38f