Introduce abstract lang string to V3
We introduce an abstract class, `Abstract_lang_string`, and many other
derived concrete classes to model the langStrings of different length,
as specified in the book.

The abstract class `Abstract_lang_string` is explicitly modelled without
the model type in order to avoid confusion for the reader of JSON and
XML schemas.