Update existing refs to the curr version of the book (#225)
We update the references in the book. Originally, the meta-model was
defined in a single self-referencing book, but eventually it was 
split in multiple parts for better readability.

We still did not reflect this in aas-core-codegen, so the references 
in this patch refer sometimes to DotAAS Part 1, but other times to 
Part 3.

Wherever possible, we referenced DotAAS Part 1a. When necessary, we 
added an inline comment when referencing DotAAS Part 3a.

In the future, we will change aas-core-codegen to operate on URLs and
deprecate structured references (such as section index).

Co-authored-by: zrgt <garmaev@gmx.de>