[V3] Note the location of `AASd-129` invariant (#218)
Some confusion was caused by the location where constraint AASd-129 is
described, and the location where it was implemented as invariant.

Namely, the constraint is described in the class `Qualifiable`, while 
the invariant belongs to the class `Submodel`. Since we can not access 
parent instances in the meta-model, we had to define the invariant at 
`Submodel` class so that we could access all the children.