[V3.0] Implement missing constraint logic for AASd-117. Fixes #253
This implements the missing invariant for AASd-117 namely, that
any `Referable` that is not direct child of a `Submodel_element_list`
must have its `ID_short` set.

Since it is impossible to access an object's parent, we need to
implement it in a creative way. Therefore, we add the invariant
to every `Referable` object that it can apply to:

- `Asset_administration_shell`: An AAS-object can never be a child
of a `Submodel_element_list`

- `Submodel`: Similarly, a `Submodel`-object is never a child of
a `Submodel_element_list`. Additionaly, every child of a `Submodel`
is evidentally not a child of a `Submodel_element_list`.

- `Submodel_element_collection`: Every child of a SEC is never a
child of a `Submodel_element_list`. Careful, though, a
`Submodel_element_collection` may be a child of a
`Submodel_element_list` though. Therefore, we cannot check for its
own `ID_short` to be set.