[V3] Fix AASc-3a-010-for-NAND (#281)
We misinterpreted the text of the constraint AASc-3a-010. It does *not*
mean an exclusive OR, but rather a NAND. Both `value` and `value_list`
*should not* be specified.

This patch adjusts the code to fit the text, to the best of our
understanding.

We also remove the redundant part of the constraint related to the
length of `value_reference_pairs` of `value_list` as that is already
verified in the class `Value_list`.

Fixes #279.