Fix htmlgen for pattern verifications (#293)
We transpile pattern verification functions to HTML code. This is
something we haven't done in codegen, and a bug eventually slipped in
where the errors were unexpectedly silenced. Notable, the `re.match`
function has not been set properly in the transpilation environment.

In this patch, we fix both the error reporting in the main program as
well as the transpilation of pattern verifications.

Fixes #292.