Specify V3 for HTTP API
We specify the data structures for V3 of the meta-model meant for HTTP
API (*i.e.*, Part 2).

In addition, we have to update to the latest aas-core-codegen in dev.
dependencies since we need to write ``result``, a reserved keyword in
Eiffel, in the spec. Consequently, we had to relax the set of the
reserved keywords in aas-core-codegen.