[V3] Fix missing literal in `FragmentKeys` (#337)
We missed to follow the book and add `Referable` literal to the constant
set `FragmentKeys`. In this patch, we complete the constant set
according to the book.

It is important to note here that the specification ("the book") is
broken. Namely, the constant set `AAS_referables_non_identifiables` *can
not* contain `Referable` literal as all identifiables are also
referables. Hence, referables + non-identifiables *can not* also contain
general referables which would be identifiables, leading to a
contradiction.

We finally decided to follow the book even though the constant set
`AAS_referables_non_identifiables` is either broken or a misnomer.

The issue has been originally discovered in:
https://github.com/aas-core-works/aas-core3.0-python/issues/37