Revert "[V3] Fix missing literal in `FragmentKeys` (#337)"
This reverts commit f41646aa4baa5807cefeb4c3bb95f7937d001b50.

Originally, we thought that we missed the literal `Referable`
in `FragmentKeys`.
In #337, we introduced that literal, together with all the
logical problems and paradoxes that come with it.
However, as it turns out, it was a bug in the specification,
rather than our mistake.
Therefore, we revert the changes here.

See [aas-core-works/aas-core3.0-python#37](https://github.com/aas-core-works/aas-core3.0-python/issues/37)
for the discussion, as it was discovered there.