[V3] Update to v3.0.1 (#334)
# Revert Embedded_data_specification Optional

In #328, we made the attribute `data_specification` in class
`Embedded_data_specification` optional, since that is what we thought
the specification said. See also the discussion in #326.

However, this was a bug in the specification, which is fixed in v3.0.1
of the specification. We therefore revert these changes.

Fixes #326.

# Remove matches_RFC_8089_path invariant from Path_type

Currently, the regex for Path type is non-compliant to the specification
v3.0.1, as it does not allow for AASX packages to be written.

As a temporary bugfix in v3.0.1, it was decided to remove the invariant
check, as changing the pattern would result in a breaking change. This
is of course no final solution, there will be a better fix for version
3.1.

See also [admin-shell-io/aas-specs#299].

[admin-shell-io/aas-specs#299]: https://github.com/admin-shell-io/aas-specs/issues/299