v3.0: Revert order of `Embedded_data_specification` attributes
In #328, we had erroneously made
`Embedded_data_specification.data_specification`
optional, changing the order of the attributes of
this class, as optional attributes have to be
specified after mandatory ones.

In #329, we reverted the `Optional`, but kept the
order of the attributes, as this didn't affect
our code. However, as a result, the generated
schema files that depend on the order of the
attributes were suddenly not backward compatible
anymore, as found in [aas-specs#477].

This reverts the order of the attributes in class
`Embedded_data_specification`.

[aas-specs#477]: https://github.com/admin-shell-io/aas-specs/issues/477