V3.1: Remove AASd-120
In v3.1, AASd-120 is relaxed from
"idShort of submodel elements being a direct
child of a SubmodelElementList shall not
be specified."
to
"idShort of submodel elements being a
direct child of a SubmodelElementList shall
not be specified within Submodels with
kind=Instance."

This makes the respective invariant not easily
implementable, considering we would need to
resolve the respective `Submodel` parent for a
given `Submodel_element_list`.

We therefore remove the invariant check and
document that we did not implement it.

See [aas-specs#432].

[aas-specs#432]: https://github.com/admin-shell-io/aas-specs/issues/432