Adapt changes for specification v3.1 (#352)
* V3.1: Adapt Identifier length to 2048

Previously, the invariant declaring the maximum
length of `Identifier` was set to `2000`. We had
already adapted this to `2024` according to
[aas-specs#306], however the specification was
corrected to `2048` in [aas-specs#306].

We therefore adapt the invariant to a maximum
length of `2048`

[aas-specs#306](https://github.com/admin-shell-io/aas-specs/issues/306)

* V3.1: Adapt Value_type_IEC_61360 length to 2048

Previously, the invariant declaring the maximum
length of `Value_type_IEC_61360` was set to
`2000`. Decided in [aas-specs#306] and specified
in [aas-specs-iec61360#21] this was adapted to
`2048`.

We therefore adapt the invariant to a maximum
length of `2048`

[aas-specs#306](https://github.com/admin-shell-io/aas-specs/issues/306)
[aas-specs-iec61360#21](https://github.com/admin-shell-io/aas-specs-iec61360/pull/21)

* V3.1: Make first and second of RelationshipElement optional

Previously, `RelationshipElement.first` and
`RelationshipElement.second` were mandatory
attributes. As decided in [aas-specs#412] and
specified in [aas-specs#431], we now make these
attributes optional.
Furthermore, we extend these changes to the
constructor of `AnnotatedRelationshipElement`.

Fixes #340

[aas-specs#412](https://github.com/admin-shell-io/aas-specs/issues/412)
[aas-specs#431](https://github.com/admin-shell-io/aas-specs/pull/431)

* V3.1: Make content_type of File and Blob optional

Previously, `File.content_type` and
`Blob.content_type` were mandatory attributes.
As decided in [aas-specs#412] and specified in
[aas-specs#431], we now make these attributes
optional.

Fixes #339

[aas-specs#412](https://github.com/admin-shell-io/aas-specs/issues/412)
[aas-specs#431](https://github.com/admin-shell-io/aas-specs/pull/431)

* v3.1: Add verifcation function for RFC 2396, adapt Path_type (#349)

Version 3.1 of the metamodel changes the
`Path_type` to be conforming to RFC 2396
(see [spec-link]).
As we did not have a verifcation function for this
RFC, we add the function `matches_RFC_2396` based
on this [abnf_to_regexp test case] and adapt the
`Path_type` invariant accordingly.

[spec-link](https://admin-shell-io.github.io/aas-specs-antora/IDTA-01001/v3.1/spec-metamodel/datatypes.html#PathType)
[abnf_to_regexp test case](https://github.com/aas-core-works/abnf-to-regexp/blob/main/test_data/nested-python/rfc2396/expected.py)

* v3.1: Adapt pattern of matches_xs_any_URI (#350)

Previously, the pattern of `matches_xs_any_URI`
followed RFC 3987, which actually specifies IRIs.

We adapt the pattern of this verification function
to follow RFC 2396, as amended by RFC 2732,
implementing the XSD v1.0 specification verbatim.

* V3.1: Remove AASd-120

In v3.1, AASd-120 is relaxed from
"idShort of submodel elements being a direct
child of a SubmodelElementList shall not
be specified."
to
"idShort of submodel elements being a
direct child of a SubmodelElementList shall
not be specified within Submodels with
kind=Instance."

This makes the respective invariant not easily
implementable, considering we would need to
resolve the respective `Submodel` parent for a
given `Submodel_element_list`.

We therefore remove the invariant check and
document that we did not implement it.

See [aas-specs#432].

[aas-specs#432]: https://github.com/admin-shell-io/aas-specs/issues/432

* V3.1: Reformulate AASd-116

In v3.1 of the specification, the
constraint AASd-116 was reformulated.
We adapt these changes.

See [aas-specs#298] and [aas-specs#390].

[aas-specs#298]: https://github.com/admin-shell-io/aas-specs/issues/298
[aas-specs#390]: https://github.com/admin-shell-io/aas-specs/issues/390

* V3.1: Remove AASd-090

Since `Referable.category` has been deprecated in
the metamodel, we remove the now unnecessary
invariant `AASd-090` about
`Data_element.category`.

See [aas-specs#514].

[aas-specs#514]: https://github.com/admin-shell-io/aas-specs/issues/514

* V3.1 Update max length of Content_type

Previously, class `Content_type` had a maximum
length of 100 characters. This was changed to 128
in v3.1 of the specification.

See:
- Issue: [aas-specs#540](https://github.com/admin-shell-io/aas-specs/pull/540)
- PR: [aas-specs#543](https://github.com/admin-shell-io/aas-specs/pull/543)
- [Specification](https://admin-shell-io.github.io/aas-specs-antora/IDTA-01001/v3.1/spec-metamodel/datatypes.html#ContentType)

* V3.1: Remove references to AASd-120 in docstrings

We previously already removed AASd-120 in #358,
but we appear to have missed two references to it
in the docstrings.
This removes the now outdated references in the
docstrings.

* V3.1: Make `Value_reference_pair.value_id` optional

Previously, `Value_reference_pair.value_id` was
mandatory. In v3.1 it became optional.

See the [changelog] (though it links to a wrong
issue).

[changelog]: https://admin-shell-io.github.io/aas-specs-antora/IDTA-01003-a/v3.1/changelog.html#_changes_v3_1_vs_v3_0_2

* v3.1: Update docstring of `ID_shorts_are_unique` (#364)

The definition of the `ID_shorts_are_unique`
invariant was updated in v3.1 of the
[specification].

This reflects the change.

[specification]: https://admin-shell-io.github.io/aas-specs-antora/IDTA-01001/v3.1/spec-metamodel/common.html#Referable

Fixes #319