Fix inheritance of class `Value_data_type` (#365)
Previously, class `Value_data_type` inherited
directly from `str`, when it should have been
`Non_empty_XML_serializable_string`, as with
every other string-class in the specification.

Since this was a bug with aas-core and not with
the specification, this adapts class inheritance
for both `v3.py` and `v3_1.py`.

Fixes #331