[V3.0 and V3.1] Fix references index constraint (#370)
AASd-128 requires, that submodel element list entries in model reference
paths are followed by an integer index in said list. In other places in
the spec, this index is described as zero-based. Currently, the
invariant for AASd-128, however, checks for the index to be a positive
integer, i.e., it excludes 0 from the set of valid values.

We instead constrain the index to non-negative integers to allow zero.