[v3.1] Revert order of Embedded_data_specification (#371)
Previously we mistakenly thought that attribute `data_specification` of
class `Embedded_data_specification` was optional and we did a bugfix for
this in #328. However, we later discovered that this was a mistake in a
graphic of the specification and `data_specification` was indeed
mandatory (see discussion in #326). Therefore we reverted the added
`Optional` in the type hint in #329.

Sadly, the issue was not finished with this, as we forgot to also revert
the order of the attributes that had changed due to the fact that
optional attributes need to be defined behind mandatory attributes.
This had been found in [aas-specs#477] with regards to V3.0 and
had been fixed for v3.0 in #353. Unfortunately we forgot to adapt this
change also for V3.1.

So now, another issue was filed in [aas-specs#585], notifying us of the
changed order of attributes in V3.1.

Finally, this reverts the order of the attributes of
`Embedded_data_specification` for V3.1 as well. This change needs to be
propagated to all schemas and SDKs.

[aas-specs#477]: https://github.com/admin-shell-io/aas-specs-metamodel/issues/477
[aas-specs#585]: https://github.com/admin-shell-io/aas-specs-metamodel/issues/585